Now able build by MinGW. All warnings were fixed. (#584)
* fix wrong libusb extract command

* update libusb to 1.0.21 under windows

* fix few -Wformat warnings

* add usleep realisation for win32

* Get rid of unused defines in mimgw.h

* fix format warning for mingw

* Add prefix SEMIXOST_ to semihost defines.
Fix redefine warnings of mingw.

+++ src/gdbserver/gdb-server.c
@@ -1428,7 +1428,7 @@ int serve(stlink_t *sl, st_state_t *st) {
 
                 reply = calloc(8 * 16 + 1, 1);
                 for(int i = 0; i < 16; i++)
-                    sprintf(&reply[i * 8], "%08x", htonl(regp.r[i]));
+                    sprintf(&reply[i * 8], "%08x", (uint32_t)htonl(regp.r[i]));
 
                 break;
